fix(tabs): show gold badge on all tabs
Move the gold indicator into the shared tabs layout so it is
available on every tab. Align the badge to the bottom-right and
remove the base-screen-only overlay implementation.
